+++ Source/SwiftLintFramework/Protocols/Rule.swift
@@ -7,11 +7,16 @@ public protocol Rule {
     init() // Rules need to be able to be initialized with default values
     init(configuration: Any) throws
 
+    func validate(file: File, compilerArguments: [String]) -> [StyleViolation]
     func validate(file: File) -> [StyleViolation]
     func isEqualTo(_ rule: Rule) -> Bool
 }
 
 extension Rule {
+    public func validate(file: File, compilerArguments: [String]) -> [StyleViolation] {
+        return validate(file: file)
+    }
+
     public func isEqualTo(_ rule: Rule) -> Bool {
         return type(of: self).description == type(of: rule).description
     }
@@ -32,10 +37,31 @@ public protocol ConfigurationProviderRule: Rule {
 }
 
 public protocol CorrectableRule: Rule {
+    func correct(file: File, compilerArguments: [String]) -> [Correction]
     func correct(file: File) -> [Correction]
 }
 
+public extension CorrectableRule {
+    func correct(file: File, compilerArguments: [String]) -> [Correction] {
+        return correct(file: file)
+    }
+}
+
 public protocol SourceKitFreeRule: Rule {}
+
+public protocol AnalyzerRule: OptInRule {}
+
+public extension AnalyzerRule {
+    func validate(file: File) -> [StyleViolation] {
+        queuedFatalError("Must call `validate(file:compilerArguments:)` for AnalyzerRule")
+    }
+}
+
+public extension AnalyzerRule where Self: CorrectableRule {
+    func correct(file: File) -> [Correction] {
+        queuedFatalError("Must call `correct(file:compilerArguments:)` for AnalyzerRule")
+    }
+}
 
 // MARK: - ConfigurationProviderRule conformance to Configurable